* split packet scheduler into three seperate parts * test against Aplle test suit, most tests pass now * suppress auxiliary records by known answers * handle very large records
git-svn-id: file:///home/lennart/svn/public/avahi/trunk@74 941a03a8-eaeb-0310-b9a0-b1bbd8fe43fe

+++ avahi-core/util.c
@@ -191,7 +191,7 @@ GTimeVal *avahi_elapse_time(GTimeVal *tv, guint msec, guint jitter) {
     return tv;
 }
 
-gint avahi_age(const GTimeVal *a) {
+glong avahi_age(const GTimeVal *a) {
     GTimeVal now;
     
     g_assert(a);
